+++ lib/td_cache/event_stream.ex
@@ -4,14 +4,13 @@ defmodule TdCache.EventStream do
   """
 
   def child_spec(config) do
-    consumer_id = config[:consumer_id]
-    consumer_group = config[:consumer_group]
+    consumer_config = Keyword.take(config, [:redis_host, :port, :consumer_id, :consumer_group])
 
     children =
       config[:streams]
       |> Enum.with_index()
       |> Enum.flat_map(fn {stream_config, i} ->
-        stream_workers(stream_config, consumer_group, consumer_id, i)
+        stream_workers(consumer_config, stream_config, i)
       end)
 
     # Spec for the supervisor that will supervise the EventStream consumers.
@@ -22,25 +21,20 @@ defmodule TdCache.EventStream do
     }
   end
 
-  defp stream_workers(stream_config, consumer_group, consumer_id, i) do
-    stream = stream_config[:key]
-    consumer = stream_config[:consumer]
-    block = Keyword.get(stream_config, :block, 1_000)
-    count = Keyword.get(stream_config, :count, 8)
-    quiesce = Keyword.get(stream_config, :quiesce, 5_000)
-    interval = Keyword.get(stream_config, :interval, 200)
-
+  defp stream_workers(consumer_config, stream_config, i) do
     [
       Supervisor.child_spec(
         {TdCache.EventStream.Consumer,
-         stream: stream,
-         consumer_group: consumer_group,
-         consumer_id: consumer_id,
-         consumer: consumer,
-         block: block,
-         count: count,
-         quiesce: quiesce,
-         interval: interval},
+         redis_host: Keyword.get(consumer_config, :redis_host, "redis"),
+         port: Keyword.get(consumer_config, :port, 6379),
+         stream: stream_config[:key],
+         consumer_group: consumer_config[:consumer_group],
+         consumer_id: consumer_config[:consumer_id],
+         consumer: stream_config[:consumer],
+         block: Keyword.get(stream_config, :block, 1_000),
+         count: Keyword.get(stream_config, :count, 8),
+         quiesce: Keyword.get(stream_config, :quiesce, 5_000),
+         interval: Keyword.get(stream_config, :interval, 200)},
         id: {EventStream.Consumer, i}
       )
     ]